+++ frontend/package.json
@@ -11,11 +11,13 @@
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
+    "@types/react-datepicker": "^4.4.2",
     "date-fns": "^2.29.2",
     "emotion-reset": "^3.0.1",
     "framer-motion": "^7.2.1",
     "next": "12.2.5",
     "react": "18.2.0",
+    "react-datepicker": "^4.8.0",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
