+++ frontend/src/components/common/Container/style.tsx
@@ -2,24 +2,25 @@ import styled from "@emotion/styled";
 import { motion } from "framer-motion";
 
 export const ImageContainer = styled(motion.div)`
-  width: 85%;
   display: flex;
+  align-items: center;
   justify-content: center;
-  position: relative;
-  margin-right: 24px;
-  top: 150px;
+
+  height: 60vh;
 `;
 
-export const MoveContainer = styled.div`
-  position: absolute;
+export const MoveContainer = styled(motion.div)`
   width: 60px;
   height: 30px;
+  margin-bottom: 22px;
   display: flex;
   justify-content: space-between;
-  top: 7%;
+
+  a {
+    margin-right: 10px;
+  }
 
   span {
-    margin-left: 10px;
     line-height: 30px;
 
     b {
