--- frontend/src/pages/missonInput.tsx
@@ -1,13 +1,0 @@
-import {BoxStyle } from '../components/common/BlueBox'
-import { StyledInput  } from "components/common/Input/styles";
-import Image from "next/image";
-import MobileLayout from "components/common/Layout/MobileLayout";
-import styled from '@emotion/styled';
-
-const VectorContainer = styled.div`
-  
-`
-
-export default function MissonInput() {
-
-}
