+++ frontend/tsconfig.json
@@ -3,7 +3,7 @@
     "jsxImportSource": "@emotion/react",
     "baseUrl": "src",
     "paths": {
-      "public/*": ["../public/*"]
+      "/public/*": ["../public/*"]
     },
     "target": "es5",
     "lib": ["dom", "dom.iterable", "esnext"],
